fix: mock 데이터 분리 (mockApi / mockData)
