+++ client/src/components/ProductList/mock/data.ts
@@ -1,7 +1,7 @@
-import MockImg from '../../assets/images/towel.png';
-import { ProductItemType, Order } from '../../types/product';
+import MockImg from '../../../assets/images/towel.png';
+import { ProductItemType } from '../../../types/product';
 
-const mockProductList: Array<ProductItemType> = [
+export const mockProductList: Array<ProductItemType> = [
   {
     id: 0,
     name: '컴퓨터싸인펜. 오늘이 전설이 될 것이다',
@@ -195,37 +195,3 @@ const mockProductList: Array<ProductItemType> = [
     imgSrc: MockImg,
   },
 ];
-
-export const apiMock = {
-  getProductList: (order: Order, page: number) => {
-    return {
-      totalProductCount: mockProductList.length,
-      totalPage: Math.floor(mockProductList.length / 20) + 1,
-      productList: getPageProducts(sortProductList(mockProductList, order), page),
-    };
-  },
-};
-
-const PRODUCT_PER_PAGE = 20;
-
-const getPageProducts = (productList: ProductItemType[], page: number) => {
-  if (page <= 0) return [];
-
-  return productList.slice(PRODUCT_PER_PAGE * (page - 1), PRODUCT_PER_PAGE * page);
-};
-
-function sortProductList(productList: Array<ProductItemType>, order: Order) {
-  switch (order) {
-    case Order.popularity:
-      return productList.sort((a: ProductItemType, b: ProductItemType) => b.point - a.point);
-    case Order.priceLow:
-      return productList.sort((a: ProductItemType, b: ProductItemType) => a.price - b.price);
-    case Order.priceHigh:
-      return productList.sort((a: ProductItemType, b: ProductItemType) => b.price - a.price);
-    case Order.recent:
-      return productList.sort(
-        (a: ProductItemType, b: ProductItemType) =>
-          new Date(b.uploadDate).getTime() - new Date(a.uploadDate).getTime()
-      );
-  }
-}
